server/requirements.txt
pyzmq==25.1.2
protobuf==4.25.3
grpcio-tools==1.62.0
